Move NEISearchHelper to separate class
This should make it  notloading NEI-Classes when ItemSearchHelper is accessed, but only if ItemSearchHelperNEI is used.
