+++ src/main/java/moze_intel/projecte/utils/ItemSearchHelper.java
@@ -1,21 +1,16 @@
 package moze_intel.projecte.utils;
 
-import codechicken.nei.ItemList;
-import codechicken.nei.SearchField;
-import codechicken.nei.api.ItemFilter;
 import cpw.mods.fml.common.Loader;
 
 import net.minecraft.item.ItemStack;
 
-import java.util.LinkedList;
-import java.util.List;
 import java.util.Locale;
 
 public abstract class ItemSearchHelper
 {
 	public static ItemSearchHelper create(String searchString) {
 		if (Loader.isModLoaded("NotEnoughItems")) {
-			return new NEISearch(searchString);
+			return new ItemSearchHelperNEI(searchString);
 		} else {
 			return new DefaultSearch(searchString);
 		}
@@ -67,39 +62,6 @@ public abstract class ItemSearchHelper
 				return false;
 			}
 			return true;
-		}
-	}
-
-	private static class NEISearch extends ItemSearchHelper
-	{
-
-		ItemFilter filter;
-		public NEISearch(String searchString)
-		{
-			super(searchString);
-			filter = getFilter(searchString);
-		}
-
-		public ItemFilter getFilter(String s_filter)
-		{
-			//based on https://github.com/Chicken-Bones/NotEnoughItems/blob/a1879a96548d17f5c4d95b40956d68f6f9db82f8/src/codechicken/nei/SearchField.java#L124-L139
-			List<ItemFilter> primary = new LinkedList<ItemFilter>();
-			List<ItemFilter> secondary = new LinkedList<ItemFilter>();
-			for (SearchField.ISearchProvider p : SearchField.searchProviders) {
-				ItemFilter filter = p.getFilter(s_filter);
-				if (filter != null)
-					(p.isPrimary() ? primary : secondary).add(filter);
-			}
-
-			if (!primary.isEmpty()) return new ItemList.AnyMultiItemFilter(primary);
-			if (!secondary.isEmpty()) return new ItemList.AnyMultiItemFilter(secondary);
-			return new ItemList.EverythingItemFilter();
-		}
-
-		@Override
-		public boolean doesItemMatchFilter_(ItemStack itemStack)
-		{
-			return filter.matches(itemStack);
 		}
 	}
 }
